// src/app/services/models/persona.ts
export class Persona {

    id_persona?: number;
    cedula: string='';
    primer_nombre: string='';
    segundo_nombre: string='';
    primer_apellido: string='';
    segundo_apellido: string='';
    genero: string='';
    fecha_nacimiento ?: string='';   
    nacionalidad: string='';   
    descripcion: string='';
    correo: string='';
    direccion: string='';
    celular: string='';
    estado: boolean;

}